+++ packages/infinity-modules/planning/src/components/onboarding/import/ReviewData.vue
@@ -128,6 +128,9 @@ export default {
         field: tag.tagName,
       }));
     },
+    save() {
+      this.$emit('save', this.rowData);
+    },
   },
 };
 </script>